meme-snipe-v18/dashboard/requirements.txt
redis>=6.2
Flask==3.1.1
requests>=2.32.3
websocket-client>=1.8.0
python-dotenv==1.0.1
humanize==4.9.0
pandas>=2.0.0
numpy>=1.24.0
